requirements.txt
click==7.1.2
country-list==0.2.1
DateTime==4.3
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.5
Werkzeug==1.0.1
psycopg2-binary==2.8.3
zope.interface==5.2.0
